robust cleanup for non-hpc test image (#3)

+++ common/clear_history.sh
@@ -64,13 +64,18 @@ fi
 # Remove Defender
 if [[ $distro == *"Ubuntu"* ]]
 then
-    apt-get purge -y mdatp
+    if dpkg -l | grep -qw mdatp; then
+        apt-get purge -y mdatp
+    fi
 elif [[ $distro == *"AzureLinux"* ]]
 then
-    tdnf remove -y mdatp
+    if tdnf list installed | grep -qw mdatp; then
+        tdnf remove -y mdatp
+    fi
 else
-    yum remove -y mdatp
-
+    if yum list installed | grep -qw mdatp; then
+        yum remove -y mdatp
+    fi
 fi
 
 # Clear History
